requirements.txt
tensorflow==2.13.0-rc2
matplotlib==3.7.1
Pillow==9.5.0
numpy==1.24.3
scikit-learn==1.2.2
tensorflow-datasets==4.9.2